refactor: 💡 use safe keyboardata load
